Restructure tests on subcycling and initialization.
