+++ src/precice/tests/SerialTests.cpp
@@ -328,117 +328,6 @@ BOOST_AUTO_TEST_CASE(testExplicitWithSubcycling)
   }
 }
 
-/// Test to run a simple coupling with subcycling.
-/// Ensures that each time step provides its own data, but preCICE will only exchange data at the end of the window.
-BOOST_AUTO_TEST_CASE(testExplicitReadWriteScalarDataWithSubcycling)
-{
-  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-
-  SolverInterface precice(context.name, _pathToTests + "explicit-scalar-data-init.xml", 0, 1); // serial coupling, SolverOne first
-
-  MeshID meshID;
-  DataID writeDataID;
-  DataID readDataID;
-
-  typedef double (*DataFunction)(double, int);
-
-  DataFunction dataOneFunction = [](double t, int idx) -> double {
-    return (double) (2 + t + idx);
-  };
-  DataFunction dataTwoFunction = [](double t, int idx) -> double {
-    return (double) (10 + t + idx);
-  };
-  DataFunction writeFunction;
-  DataFunction readFunction;
-
-  if (context.isNamed("SolverOne")) {
-    meshID        = precice.getMeshID("MeshOne");
-    writeDataID   = precice.getDataID("DataOne", meshID);
-    writeFunction = dataOneFunction;
-    readDataID    = precice.getDataID("DataTwo", meshID);
-    readFunction  = dataTwoFunction;
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    meshID        = precice.getMeshID("MeshTwo");
-    writeDataID   = precice.getDataID("DataTwo", meshID);
-    writeFunction = dataTwoFunction;
-    readDataID    = precice.getDataID("DataOne", meshID);
-    readFunction  = dataOneFunction;
-  }
-
-  int n_vertices = 1;
-
-  std::vector<VertexID> vertexIDs(n_vertices, 0);
-  std::vector<double>   writeData(n_vertices, 0);
-  std::vector<double>   readData(n_vertices, 0);
-  double                oldWriteData, oldReadData;
-
-  vertexIDs[0] = precice.setMeshVertex(meshID, Eigen::Vector3d(0.0, 0.0, 0.0).data());
-
-  int    nSubsteps     = 4; // perform subcycling on solvers. 4 steps happen in each window.
-  int    nWindows      = 5; // perform 5 windows.
-  double maxDt         = precice.initialize();
-  double windowDt      = maxDt;
-  int    timestep      = 0;
-  int    timewindow    = 0;
-  double dt            = windowDt / (nSubsteps - 0.5); // Timestep length desired by solver. E.g. 4 steps with size 2/7. Fourth step will be restricted to 1/7 via preCICE steering to fit into the window.
-  double expectedDts[] = {2.0 / 7.0, 2.0 / 7.0, 2.0 / 7.0, 1.0 / 7.0};
-  double currentDt     = dt; // Timestep length used by solver
-  double time          = timestep * dt;
-
-  if (precice.isActionRequired(precice::constants::actionWriteInitialData())) {
-    for (int i = 0; i < n_vertices; i++) {
-      writeData[i] = writeFunction(time, i);
-      precice.writeScalarData(writeDataID, vertexIDs[i], writeData[i]);
-    }
-    precice.markActionFulfilled(precice::constants::actionWriteInitialData());
-  }
-
-  precice.initializeData();
-
-  while (precice.isCouplingOngoing()) {
-    double readTime;
-    if (context.isNamed("SolverOne")) {
-      readTime = timewindow * windowDt; // SolverOne lags one window behind SolverTwo for serial-explicit coupling.
-    } else {
-      readTime = (timewindow + 1) * windowDt;
-    }
-    BOOST_TEST(readData.size() == n_vertices);
-    for (int i = 0; i < n_vertices; i++) {
-      oldReadData = readData[i];
-      precice.readScalarData(readDataID, vertexIDs[i], readData[i]);
-      if (precice.isTimeWindowComplete() ||
-          (timestep == 0)) {                      // exception: First timestep will also have different data, even though formally no time window is completed.
-        BOOST_TEST((readData[i] != oldReadData)); // ensure that read data changes from one step to the next, if a new window is entered
-      } else {
-        BOOST_TEST((readData[i] == oldReadData)); // ensure that read data stays the same from one step to the next, if not a new window is entered
-      }
-      BOOST_TEST(readData[i] == readFunction(readTime, i));
-    }
-
-    // solve usually goes here. Dummy solve: Just sampling the writeFunction.
-    BOOST_TEST(currentDt == expectedDts[timestep % nSubsteps]);
-    time += currentDt;
-
-    BOOST_TEST(writeData.size() == n_vertices);
-    for (int i = 0; i < n_vertices; i++) {
-      oldWriteData = writeData[i];
-      writeData[i] = writeFunction(time, i);
-      BOOST_TEST(writeData[i] != oldWriteData); // ensure that write data differs from one step to the next
-      precice.writeScalarData(writeDataID, vertexIDs[i], writeData[i]);
-    }
-    maxDt     = precice.advance(currentDt);
-    currentDt = dt > maxDt ? maxDt : dt;
-    timestep++;
-    if (precice.isTimeWindowComplete()) {
-      timewindow++;
-    }
-  }
-
-  precice.finalize();
-  BOOST_TEST(timestep == nWindows * nSubsteps);
-}
-
 /// One solver uses incremental position set, read/write methods.
 /// @todo This test uses resetmesh. How did this ever work?
 #if 0
@@ -532,63 +421,6 @@ BOOST_AUTO_TEST_CASE(testExplicitWithDataExchange)
   }
 }
 #endif
-
-/**
- * @brief The second solver initializes the data of the first.
- *
- * A mapping is employed for the second solver, i.e., at the end of
- * initializeData(), the mapping needs to be invoked.
- */
-BOOST_AUTO_TEST_CASE(testExplicitWithDataInitialization)
-{
-  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-
-  using Eigen::Vector3d;
-
-  SolverInterface cplInterface(context.name, _pathToTests + "explicit-data-init.xml", 0, 1);
-  if (context.isNamed("SolverOne")) {
-    int meshOneID = cplInterface.getMeshID("MeshOne");
-    cplInterface.setMeshVertex(meshOneID, Vector3d(1.0, 2.0, 3.0).data());
-    double maxDt      = cplInterface.initialize();
-    int    dataAID    = cplInterface.getDataID("DataOne", meshOneID);
-    int    dataBID    = cplInterface.getDataID("DataTwo", meshOneID);
-    double valueDataB = 0.0;
-    cplInterface.initializeData();
-    cplInterface.readScalarData(dataBID, 0, valueDataB);
-    BOOST_TEST(2.0 == valueDataB);
-    while (cplInterface.isCouplingOngoing()) {
-      Vector3d valueDataA(1.0, 1.0, 1.0);
-      cplInterface.writeVectorData(dataAID, 0, valueDataA.data());
-      maxDt = cplInterface.advance(maxDt);
-      cplInterface.readScalarData(dataBID, 0, valueDataB);
-      BOOST_TEST(2.5 == valueDataB);
-    }
-    cplInterface.finalize();
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    int      meshTwoID = cplInterface.getMeshID("MeshTwo");
-    Vector3d pos       = Vector3d::Zero();
-    cplInterface.setMeshVertex(meshTwoID, pos.data());
-    double maxDt   = cplInterface.initialize();
-    int    dataAID = cplInterface.getDataID("DataOne", meshTwoID);
-    int    dataBID = cplInterface.getDataID("DataTwo", meshTwoID);
-    cplInterface.writeScalarData(dataBID, 0, 2.0);
-    //tell preCICE that data has been written and call initializeData
-    cplInterface.markActionFulfilled(precice::constants::actionWriteInitialData());
-    cplInterface.initializeData();
-    Vector3d valueDataA;
-    cplInterface.readVectorData(dataAID, 0, valueDataA.data());
-    Vector3d expected(1.0, 1.0, 1.0);
-    BOOST_TEST(valueDataA == expected);
-    while (cplInterface.isCouplingOngoing()) {
-      cplInterface.writeScalarData(dataBID, 0, 2.5);
-      maxDt = cplInterface.advance(maxDt);
-      cplInterface.readVectorData(dataAID, 0, valueDataA.data());
-      BOOST_TEST(valueDataA == expected);
-    }
-    cplInterface.finalize();
-  }
-}
 
 /// One solver uses block set/get/read/write methods.
 /// @todo This test uses resetmesh. How did this ever work?
@@ -1289,71 +1121,6 @@ BOOST_AUTO_TEST_CASE(testImplicit)
     couplingInterface.finalize();
     BOOST_TEST(computedTimesteps == 4);
   }
-}
-
-/// Test simple coupled simulation with iterations, data initialization and without acceleration
-BOOST_AUTO_TEST_CASE(testImplicitWithDataInitialization)
-{
-  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
-
-  using namespace precice::constants;
-
-  SolverInterface couplingInterface(context.name, _pathToTests + "implicit-data-init.xml", 0, 1);
-
-  int         dimensions = couplingInterface.getDimensions();
-  std::string meshName;
-  std::string writeDataName;
-  std::string readDataName;
-  double      writeValue, expectedReadValue;
-
-  if (context.isNamed("SolverOne")) {
-    meshName          = "MeshOne";
-    writeDataName     = "Forces";
-    readDataName      = "Velocities";
-    writeValue        = 1;
-    expectedReadValue = 2;
-  } else {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    meshName          = "MeshTwo";
-    writeDataName     = "Velocities";
-    readDataName      = "Forces";
-    writeValue        = 2;
-    expectedReadValue = 1;
-  }
-  int                 meshID      = couplingInterface.getMeshID(meshName);
-  int                 writeDataID = couplingInterface.getDataID(writeDataName, meshID);
-  int                 readDataID  = couplingInterface.getDataID(readDataName, meshID);
-  std::vector<double> vertex(dimensions, 0);
-  int                 vertexID = couplingInterface.setMeshVertex(meshID, vertex.data());
-
-  double dt = 0;
-  dt        = couplingInterface.initialize();
-  std::vector<double> writeData(dimensions, writeValue);
-  std::vector<double> readData(dimensions, -1);
-  const std::string & cowid = actionWriteInitialData();
-
-  if (couplingInterface.isActionRequired(cowid)) {
-    BOOST_TEST(context.isNamed("SolverTwo"));
-    couplingInterface.writeVectorData(writeDataID, vertexID, writeData.data());
-    couplingInterface.markActionFulfilled(cowid);
-  }
-
-  couplingInterface.initializeData();
-
-  while (couplingInterface.isCouplingOngoing()) {
-    if (couplingInterface.isActionRequired(actionWriteIterationCheckpoint())) {
-      couplingInterface.markActionFulfilled(actionWriteIterationCheckpoint());
-    }
-    couplingInterface.readVectorData(readDataID, vertexID, readData.data());
-    BOOST_TEST(expectedReadValue == readData.at(0));
-    BOOST_TEST(expectedReadValue == readData.at(1));
-    couplingInterface.writeVectorData(writeDataID, vertexID, writeData.data());
-    dt = couplingInterface.advance(dt);
-    if (couplingInterface.isActionRequired(actionReadIterationCheckpoint())) {
-      couplingInterface.markActionFulfilled(actionReadIterationCheckpoint());
-    }
-  }
-  couplingInterface.finalize();
 }
 
 /// Tests stationary mapping with solver provided meshes.
