+++ src/precice/tests/implicit-scalar-data-init.xml
@@ -39,12 +39,14 @@
 
     <m2n:sockets from="SolverOne" to="SolverTwo" />
 
-    <coupling-scheme:serial-explicit>
+    <coupling-scheme:serial-implicit>
       <participants first="SolverOne" second="SolverTwo" />
       <max-time-windows value="5" />
       <time-window-size value="2" />
+      <max-iterations value="3" />
+      <min-iteration-convergence-measure min-iterations="3" data="DataOne" mesh="MeshOne" />
       <exchange data="DataOne" mesh="MeshOne" from="SolverOne" to="SolverTwo" />
       <exchange data="DataTwo" mesh="MeshOne" from="SolverTwo" to="SolverOne" initialize="on" />
-    </coupling-scheme:serial-explicit>
+    </coupling-scheme:serial-implicit>
   </solver-interface>
 </precice-configuration>
