+++ src/precice/tests/oneway-data-init-read-mapping.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<precice-configuration>
+  <solver-interface dimensions="3">
+    <data:scalar name="Data" />
+
+    <mesh name="MeshOne">
+      <use-data name="Data" />
+    </mesh>
+
+    <mesh name="MeshTwo">
+      <use-data name="Data" />
+    </mesh>
+
+    <participant name="SolverOne">
+      <use-mesh name="MeshOne" provide="on" />
+      <use-mesh name="MeshTwo" from="SolverTwo" />
+      <mapping:nearest-neighbor
+        direction="read"
+        from="MeshTwo"
+        to="MeshOne"
+        constraint="conservative"
+        timing="initial" />
+      <read-data name="Data" mesh="MeshOne" />
+    </participant>
+
+    <participant name="SolverTwo">
+      <use-mesh name="MeshTwo" provide="on" />
+      <write-data name="Data" mesh="MeshTwo" />
+    </participant>
+
+    <m2n:sockets from="SolverOne" to="SolverTwo" />
+
+    <coupling-scheme:serial-explicit>
+      <participants first="SolverOne" second="SolverTwo" />
+      <max-time-windows value="5" />
+      <time-window-size value="1.0" />
+      <exchange data="Data" mesh="MeshTwo" from="SolverTwo" to="SolverOne" initialize="on" />
+    </coupling-scheme:serial-explicit>
+  </solver-interface>
+</precice-configuration>
